♻️ refactor: converter report 용으로 분리
